sunxi: sync device tree with Linux 4.11
As the device tree of Linux have changed a lot since the initial ones
are merged into U-Boot, update it to the version shipped with Linux 4.11 .

The sun8i-emac part is manually added into the device tree, in order to
satisfy the driver in U-Boot. The pinctrl part of EMAC pins are still in
old binding as sun8i-emac doesn't recognize the new binding.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>